norminette

+++ coders/parser.h
@@ -15,14 +15,14 @@
 
 typedef struct s_config
 {
-	int		number_of_coders;
-	int		time_to_burnout;
-	int		time_to_compile;
-	int		time_to_debug;
-	int		time_to_refactor;
-	int		number_of_compiles_required;
-	int		dongle_cooldown;
-	char	*scheduler;
+	int			number_of_coders;
+	long		time_to_burnout;
+	long		time_to_compile;
+	long		time_to_debug;
+	long		time_to_refactor;
+	long		number_of_compiles_required;
+	long		dongle_cooldown;
+	char		*scheduler;
 }			t_config;
 
 bool		parse_arguments(int argc, char **argv, t_config *config);
